Corrigindo o processo de build

+++ publish.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+yarn encore prod
+if [ $? -ne 0 ]; then echo "Could not build the assets"; exit 1; fi
+
 docker run --rm -itv $(pwd):/app -u $(id -u):$(id -g) -w /app php-blog php ./vendor/bin/sculpin generate --env=prod
 if [ $? -ne 0 ]; then echo "Could not generate the site"; exit 1; fi
